Changed the max number of tokens in a place to 999 (more looks graphically ugly and we will never ever need more than this).

+++ src/pipe/gui/CreateGui.java
@@ -26,7 +26,7 @@ public class CreateGui {
 
 	public static String imgPath, userPath; // useful for stuff
 	
-	public static Integer MaximalNumberOfTokensAllowed = new Integer(9999);
+	public static Integer MaximalNumberOfTokensAllowed = new Integer(999);
 
 	public static void init() {
 		imgPath = "resources/Images/";
